Reorganised the folder structure in master
