+++ commonMethods/commonFunctions.js
