+++ commonMethods/constants.js
