+++ commonMethods/webKitPerformanceCommonFunctions.js
